Core/Scripts : Warning Build (90% complete)
- Resolved double-to-float truncation in complex math for boss_prince_taldaram (Heroic sphere trajectory logic).
- Completed the audit of MoveJump calls, ensuring all 16+ scripts use explicit boolean parameters (0.0f/1.0f).
- Validated all previous pending fixes for Northrend, Pandaria, Draenor, and Legion boss scripts.
- Standardized engine core files (Unit, Player, SpellEffects) to eliminate signed/unsigned mismatches and risky enum-to-float operations.
- Cleaned up miscellaneous infrastructure files including DBUpdater, PlayerDump, and TransportMgr.

+++ src/server/game/Battlefield/Battlefield.cpp
@@ -1097,7 +1097,7 @@ bool BfCapturePoint::Update(uint32 diff)
                 HandlePlayerEnter(*itr);
 
     // get the difference of numbers
-    float fact_diff = ((float)m_activePlayers[0].size() - (float)m_activePlayers[1].size()) * diff / BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL;
+    float fact_diff = ((float)m_activePlayers[0].size() - (float)m_activePlayers[1].size()) * diff / static_cast<float>(BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL);
     if (!fact_diff)
         return false;
